+++ packages/css/src/index.ts
@@ -12,6 +12,7 @@ export * from './identifier';
 export * from './theme';
 export * from './style';
 export * from './vars';
+export * from './layer';
 export { createContainer } from './container';
 
 export const css$ = <T>(t: T): T => t;