+++ test-helpers/src/startFixture/types.ts
@@ -1,8 +1,11 @@
-export type BuildType =
+type BuildType =
   | 'browser'
   | 'mini-css-extract'
   | 'style-loader'
   | 'esbuild'
+  | 'esbuild-next'
+  | 'esbuild-runtime'
+  | 'esbuild-next-runtime'
   | 'vite'
   | 'parcel';